requirements.txt
slack-sdk
apscheduler
requests
python-dotenv
pytz
